% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex.R
\name{apex_connect}
\alias{apex_connect}
\title{APEX Connect}
\usage{
apex_connect(user = Sys.getenv("APEX.SLU.EDU_USER"),
  pass = Sys.getenv("APEX.SLU.EDU_PASS"), verbose = FALSE)
}
\arguments{
\item{user}{apex username as string}

\item{pass}{either a string or a callback function for password prompt}

\item{verbose}{either TRUE/FALSE or a value between 0 and 4 indicating log level: 0: no logging, 1: only warnings, 2: protocol, 3: packets or 4: full stack trace.}
}
\value{
ssh session
}
\description{
APEX Connect
}
\examples{
\dontrun{
# retrive user and pass from .Renviron file in user home directory
session <- apex_connect()
}

}
\seealso{
\code{\link[ssh]{ssh_connect}}
}
