% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex.R
\name{apex_disconnect}
\alias{apex_disconnect}
\title{APEX Disconnect}
\usage{
apex_disconnect(session)
}
\arguments{
\item{session}{ssh connection created with \code{\link{apex_connect}}}
}
\description{
APEX Disconnect
}
\examples{
\dontrun{
session <- apex_connect()
apex_disconnect(session)
}

}
\seealso{
\code{\link[ssh]{ssh_disconnect}}
}
