% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex.R
\name{apex_download}
\alias{apex_download}
\title{APEX Download}
\usage{
apex_download(session, files, to = ".", verbose = FALSE)
}
\arguments{
\item{session}{ssh connection created with \code{\link{apex_connect}}}

\item{files}{path to files or directory to transfer}

\item{to}{existing directory on the destination where files will be copied into}

\item{verbose}{print progress while copying files}
}
\description{
APEX Download
}
\examples{
\dontrun{
# recursively download files and directories
session <- apex_connect()
apex_download(session, "~/target/*", tempdir())
}

}
\seealso{
\code{\link[ssh]{scp_download}}
}
