% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex.R
\name{apex_execute}
\alias{apex_execute}
\title{APEX Execute}
\usage{
apex_execute(session, command = "whoami", error = TRUE)
}
\arguments{
\item{session}{ssh connection created with \code{\link{apex_connect}}}

\item{command}{The command or script to execute}

\item{error}{automatically raise an error if the exit status is non-zero}
}
\value{
list containing exit status, buffered raw stdout, and buffered raw stderr
}
\description{
APEX Execute
}
\examples{
\dontrun{
session <- apex_connect()
out <- apex_execute(session)
rawToChar(out$stdout)
}

}
\seealso{
\code{\link[ssh]{ssh_exec_internal}}
}
