% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex.R
\name{apex_upload}
\alias{apex_upload}
\title{APEX Upload}
\usage{
apex_upload(session, files, to = "Documents", verbose = FALSE)
}
\arguments{
\item{session}{ssh connection created with \code{\link{apex_connect}}}

\item{files}{path to files or directory to transfer}

\item{to}{existing directory on the destination where files will be copied into}

\item{verbose}{print progress while copying files}
}
\description{
APEX Upload
}
\examples{
\dontrun{
# recursively upload files and directories
session <- apex_connect()
files <- c(R.home("doc"), R.home("COPYING"))
apex_upload(session, files, "~/target")
}

}
\seealso{
\code{\link[ssh]{scp_upload}}
}
