% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{slurm_cancel}
\alias{slurm_cancel}
\title{SLURM Cancel}
\usage{
slurm_cancel(session, slurm_job)
}
\arguments{
\item{session}{ssh connection created with \code{\link{apex_connect}}}

\item{slurm_job}{a slurm_job object created with \code{\link{slurm_apply}}}
}
\value{
character representation of stdout
}
\description{
SLURM Cancel
}
\examples{
session <- apex_connect()
do_search <- function(x, y) x + y
slurm_job <- slurm_apply(do_search,
                         data.frame(x = seq( 0, 42,  21),
                                    y = seq(42,  0, -21)),
                         "find_meaning")
slurm_upload(session, slurm_job)
slurm_submit(session, slurm_job)
if(slurm_is_running(session, slurm_job))
  slurm_cancel(session, slurm_job)

}
\seealso{
\code{\link{apex_execute}}
}
