% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{slurm_output_dfr}
\alias{slurm_output_dfr}
\title{SLURM Read executed job output into data frame}
\usage{
slurm_output_dfr(slurm_job)
}
\arguments{
\item{slurm_job}{a slurm_job object created with \code{\link{slurm_apply}}}
}
\value{
a data frame with one column by return value of the function passed to \code{\link{slurm_apply}}, where each row is the output of the corresponding row in the params data frame passed to \code{\link{slurm_apply}}
}
\description{
SLURM Read executed job output into data frame
}
\examples{
\dontrun{
session <- apex_connect()
do_search <- function(x, y) x + y
slurm_job <- slurm_apply(do_search,
                         data.frame(x = seq( 0, 42,  21),
                                    y = seq(42,  0, -21)),
                         "find_meaning")
slurm_upload(session, slurm_job)
slurm_submit(session, slurm_job)
slurm_download(session, slurm_job)
results <- slurm_output_dfr(slurm_job)
}

}
\seealso{
\code{\link[rslurm]{get_slurm_out}}
}
