% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{slurm_remove_apex}
\alias{slurm_remove_apex}
\title{SLURM Remove Job Directory on APEX}
\usage{
slurm_remove_apex(session, slurm_job, parent_directory = "Documents")
}
\arguments{
\item{session}{ssh connection created with \code{\link{apex_connect}}}

\item{slurm_job}{a slurm_job object created with \code{\link{slurm_apply}}}

\item{parent_directory}{location of slurm job folder on apex}
}
\value{
character representation of stdout
}
\description{
SLURM Remove Job Directory on APEX
}
\examples{
\dontrun{
session <- apex_connect()
do_search <- function(x, y) x + y
slurm_job <- slurm_apply(do_search,
                         data.frame(x = seq( 0, 42,  21),
                                    y = seq(42,  0, -21)),
                         "find_meaning")
slurm_upload(session, slurm_job)
slurm_submit(session, slurm_job)
while(slurm_is_running(session, slurm_job)) {
  Sys.sleep(1)
}
slurm_download(session, slurm_job)
results <- slurm_output_dfr(slurm_job)
slurm_remove_apex(session, slurm_job)
apex_disconnect(session)
}

}
\seealso{
\code{\link{apex_execute}}
}
