% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{slurm_remove_local}
\alias{slurm_remove_local}
\title{SLURM Remove Local Job Directory}
\usage{
slurm_remove_local(slurm_job)
}
\arguments{
\item{slurm_job}{a slurm_job object created with \code{\link{slurm_apply}}}
}
\description{
SLURM Remove Local Job Directory
}
\examples{
\dontrun{
session <- apex_connect()
do_search <- function(x, y) x + y
slurm_job <- slurm_apply(do_search,
                         data.frame(x = seq( 0, 42,  21),
                                    y = seq(42,  0, -21)),
                         "find_meaning")
slurm_upload(session, slurm_job)
slurm_submit(session, slurm_job)
slurm_download(session, slurm_job)
results <- slurm_output_dfr(slurm_job)
slurm_remove_apex(session, slurm_job)
apex_disconnect(session)
slurm_remove_local(slurm_job)
}

}
\seealso{
\code{\link[rslurm]{cleanup_files}}
}
