% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{slurm_upload}
\alias{slurm_upload}
\title{SLURM Upload}
\usage{
slurm_upload(session, slurm_job, ...)
}
\arguments{
\item{session}{ssh connection created with \code{\link{apex_connect}}}

\item{slurm_job}{a slurm_job object created with \code{\link{slurm_apply}}}

\item{...}{further arguments passed to \code{\link{apex_upload}}}
}
\value{
target location as string
}
\description{
SLURM Upload
}
\examples{
session <- apex_connect()
do_search <- function(x, y) x + y
slurm_job <- slurm_apply(do_search,
                         data.frame(x = seq( 0, 42,  21),
                                    y = seq(42,  0, -21)),
                         "find_meaning")
slurm_upload(session, slurm_job)

}
\seealso{
\code{\link{apex_upload}}
}
