% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{slurm_download}
\alias{slurm_download}
\title{SLURM Download}
\usage{
slurm_download(session, slurm_job, parent_directory = "Documents",
  seconds = 10, ...)
}
\arguments{
\item{session}{ssh connection created with \code{\link{apex_connect}}}

\item{slurm_job}{a slurm_job object created with \code{\link{slurm_apply}}}

\item{parent_directory}{location of slurm job folder on apex}

\item{seconds}{seconds between job completion checks}

\item{...}{further arguments passed to \code{\link{apex_download}}}
}
\description{
SLURM Download
}
\note{
This function will block until the job finishes on the cluster.
}
\examples{
\dontrun{
session <- apex_connect()
do_search <- function(x, y) x + y
slurm_job <- slurm_apply(do_search,
                         data.frame(x = seq( 0, 42,  21),
                                    y = seq(42,  0, -21)),
                         "find_meaning")
slurm_upload(session, slurm_job)
slurm_submit(session, slurm_job)
slurm_download(session, slurm_job)
}

}
\seealso{
\code{\link{apex_download}}
}
